GameStateController now calls reset
